fix(Fixes https://github.com/Joystream/hydra/issues/519): fixes hydra-processor syncing issue occuri
affects: @joystream/hydra-cli, @joystream/hydra-processor

\`blocksWithEvents\` \`AsyncGenerator\` was not properly awaiting the promise returned by
\`pWaitFor\`.
This promise was supposed to stop the generator from yeilding next block
until all
evens emitted in that block were fetched from the indexer.
Becasue it was not properly awaited,
it
was possible for the \`blocksWithEvents\` to yeild a block with an incomplete array of events
in case
the indexer was responsing to requests more slowly than usual.
As a result, the processor would
think it processed a full block
and update the \`last_scanned_block\` value in the database,
while not
actually processing all events in that block.
This is the scenario that has caused issues such as
https://github.com/Joystream/hydra/issues/519

ISSUES CLOSED: #519